Rename shared header to MemoryAccessInfo
