+++ core/MemoryAccessInfo.hpp
